pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gsheet-tui"
version = "0.1.0"
description = "A terminal user interface for editing Google Sheets"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth>=2.22.0",
    "gspread>=5.10.0",
    "textual>=0.40.0",
    "google-api-python-client>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src/gsheet_tui"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true

[project.scripts]
gsheet = "gsheet_tui.main:main" 